some smol changes
